feat!: Restructure the Emacs configuration and make it work
Obsoleted extra configuration files for extras.org and compat.el.
Now emacs-config.org is the single public configuration file.
Packages with extra dependencies are tagged @extra, and old settings
are enclosed in :status blocks.

There is a new sandboxing-enabled Emacs package in the overlay.
Support for custom-file is dropped. Use extraInitText instead.

A new home-manager module for an unsafe Emacs profile has been added,
and it is available by default on desktop machines.

The sandboxing script is available as a package in the overlay.