lisp/treesit: Limit the bound to eol
If the point is at the middle of an element, don't kill text after the eol, even
if the element continues after the eol.

There can be an element that spans multiple lines, e.g. multi-line/block
comment. If the point is in the middle of a block comment, a natural behavior is
to kill from the point to the end of the line.