lisp/tailscale: Fix the destination host name
There can be situations where the host name from the status output cannot be
used as the target of the copy command. For example, if the host name contains
spaces or uppercase letters, an escaped name should be used instead. The
accepted target name is available as the first component of the DNSName value.